requirements.txt
whisperx==3.1.1
faster-whisper==1.0.3
torch>=2.0.0
torchaudio>=2.0.0
numpy<2.0
pandas>=2.0.0
transformers>=4.30.0
nltk>=3.8.1
pyannote.audio>=3.1.0
flask>=2.0.0
python-dotenv>=1.0.0
lightning-whisper-mlx>=0.0.10 
psutil>=7.0.0 